refactor: separate geoip logic
